+++ homepage/views.py
@@ -1,5 +1,4 @@
 # from django.shortcuts import render  # noqa: F401  # Ignore "imported but unused"
-from django.http import HttpResponse
 from django.http import JsonResponse
 
 from django.db import connection
