+++ templates/400.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block content %}
+  <div class="d-flex flex-column justify-content-center align-items-center text-center" style="flex: 1;">
+    <h1 class="display-4 text-danger">400 - Bad Request</h1>
+    <p class="lead">Oops! Something went wrong with your request.</p>
+    <a class="btn btn-primary mt-3" href="{% url 'home' %}">Back to Home</a>
+  </div>
+{% endblock %}
